--- ReadersAndroidJetpack/app/src/main/java/com/kontinua/readersandroidjetpack/HttpClient.kt
@@ -1,10 +1,0 @@
-package com.kontinua.readersandroidjetpack
-
-import okhttp3.OkHttpClient
-
-// TODO: Remove this when API manager is fully implemented.
-object HttpClient {
-    val instance: OkHttpClient by lazy {
-        OkHttpClient()
-    }
-}
