+++ ReadersAndroidJetpack/app/src/main/java/com/kontinua/readersandroidjetpack/util/APIManager.kt
@@ -1,5 +1,6 @@
-package com.kontinua.readersandroidjetpack
+package com.kontinua.readersandroidjetpack.util
 
+import android.util.Log
 import com.kontinua.readersandroidjetpack.serialization.Collection
 import com.kontinua.readersandroidjetpack.serialization.CollectionPreview
 import com.kontinua.readersandroidjetpack.serialization.Workbook
@@ -10,23 +11,32 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import okhttp3.OkHttpClient
 import okhttp3.Request
+import java.io.File
+import java.io.FileOutputStream
 
 object APIManager {
+    /*
+    Manages any requests pertaining to the backend Kontinua Readers Service
+    i.e. fetching workbooks/collections and downloading pdfs.
+     */
 
     private const val API_URL = "http://18.189.208.93/api"
-    private val client = OkHttpClient()
-    private val moshi = Moshi.Builder().build()
+    private final val CLIENT = OkHttpClient()
+    private final val MOSHI = Moshi.Builder().build()
 
     suspend fun getCollections(localization: String = "en-US"): List<CollectionPreview>? {
+        /*
+        Fetches all collections for a given localization.
+         */
         val type = Types.newParameterizedType(List::class.java, CollectionPreview::class.java)
-        val adapter = moshi.adapter<List<CollectionPreview>>(type)
+        val adapter = MOSHI.adapter<List<CollectionPreview>>(type)
 
         return withContext(Dispatchers.IO) {
 
             val request =
                 Request.Builder().url("$API_URL/collections?localization=$localization").build()
 
-            client.newCall(request).execute().use { response ->
+            CLIENT.newCall(request).execute().use { response ->
                 if (!response.isSuccessful) throw Exception("Request failed: ${response.code}")
                 return@withContext response.body?.string()?.let { adapter.fromJson(it) }
             }
@@ -36,16 +46,16 @@ object APIManager {
 
     suspend fun getCollection(preview: CollectionPreview): Collection? {
         /*
-        Gets detailed collection given its preview
+        Given a preview, fetches the collection which will include which workbooks it offers.
          */
-        val adapter = moshi.adapter(Collection::class.java)
+        val adapter = MOSHI.adapter(Collection::class.java)
 
         val id = preview.id
 
         val request = Request.Builder().url("$API_URL/collections/$id").build()
 
         return withContext(Dispatchers.IO) {
-            client.newCall(request).execute().use { response ->
+            CLIENT.newCall(request).execute().use { response ->
                 if (!response.isSuccessful) throw Exception("Request failed: ${response.code}")
                 return@withContext response.body?.string()?.let { adapter.fromJson(it) }
             }
@@ -68,20 +78,50 @@ object APIManager {
 
     suspend fun getWorkbook(preview: WorkbookPreview): Workbook? {
 
-        val adapter = moshi.adapter(Workbook::class.java)
+        val adapter = MOSHI.adapter(Workbook::class.java)
 
         val id = preview.id
 
         val request = Request.Builder().url("$API_URL/workbooks/$id").build()
 
         return withContext(Dispatchers.IO) {
-            client.newCall(request).execute().use { response ->
+            CLIENT.newCall(request).execute().use { response ->
                 if (!response.isSuccessful) throw Exception("Request failed: ${response.code}")
 
                 return@withContext response.body?.string()?.let { adapter.fromJson(it) }
             }
 
 
+        }
+    }
+
+    suspend fun getPDFFromWorkbook(context: android.content.Context, workbook: Workbook): File? {
+        /*
+        Given a workbook, downloads PDF and returns it s a file.
+         */
+        return withContext(Dispatchers.IO) {
+            try {
+                val request = Request.Builder().url(workbook.pdf).build()
+                val response = CLIENT.newCall(request).execute()
+
+                if (!response.isSuccessful) {
+                    Log.e("Download PDF", "Failed: ${response.code}")
+                    return@withContext null
+                }
+
+                response.body?.let { body ->
+                    val file = File(context.cacheDir, "downloaded.pdf")
+                    val fos = FileOutputStream(file)
+                    fos.use { output ->
+                        output.write(body.bytes())
+                    }
+                    return@withContext file
+                }
+            } catch (e: Exception) {
+                Log.e("Download PDF", "Error downloading PDF: ${e.message}")
+                e.printStackTrace()
+                null
+            }
         }
     }
